created new models folder for better structure and moved form validation to utils, while renaming import packages
