+++ app/__init__.py
@@ -35,7 +35,7 @@ def create_app(config_object=DevConfig):
     login_manager.login_view = 'auth.login'
     login_manager.init_app(app)
 
-    from .models import User
+    from .models.models import User
     @login_manager.user_loader
     def load_user(user_id):
         return User.query.get(user_id)
