+++ app/views/auth.py
@@ -3,9 +3,9 @@ from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import login_user, logout_user, login_required, current_user
 
 
-from ..models import User
+from ..models.models import User
 from .. import db
-from ..form_validation import ChangeSite, ChangeBuilding, ChangeRoom, ChangePassword, ChangeEmail
+from ..util.form_validation import ChangeSite, ChangeBuilding, ChangeRoom, ChangePassword, ChangeEmail
 
 auth = Blueprint('auth', __name__)
